requirements.txt
selenium==4.26.1
pandas==2.2.3
openpyxl==3.1.5
tqdm==4.67.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
python-dotenv==1.0.1
webdriver-manager==4.0.2
colorama==0.4.6
numpy==1.26.4
hashlib3==2.5.0